Phase 4 Complete: API Layer Refactor
✅ API STRUCTURE CREATED:
- Created /api/ package with modular organization
- density.py, flow.py, map.py, report.py endpoints
- /api/models/ package with Pydantic schemas

✅ ROUTES RELOCATED:
- app/density_api.py → api/density.py
- app/map_api.py → api/map.py
- app/routes/api_flow.py → api/flow.py
- app/report.py → api/report.py

✅ TYPED MODELS IMPLEMENTED:
- api/models/density.py - DensityAnalysisRequest/Response
- api/models/flow.py - FlowAnalysisRequest/Response
- api/models/map.py - MapManifestResponse, MapBinsRequest/Response
- api/models/report.py - ReportRequest/Response with camelCase aliases

✅ ADAPTER LAYER IMPLEMENTED:
- All original files contain import shims
- Zero breaking changes - 100% backward compatibility
- Updated imports to use core/ modules instead of inline logic

✅ TESTING VALIDATED:
- E2E tests: All pass ✅
- Import verification: All adapter shims work correctly
- API endpoints: All responding correctly

✅ DOCUMENTATION CREATED:
- docs/PHASE4_RESULTS.md with complete analysis
- Clear separation of API routes from core business logic

Resolves: Issue #345
