+++ artifacts/2025-10-26/ui/health.json
@@ -4,38 +4,38 @@
     "platform": "Local",
     "version": "unknown",
     "data_root": "./data",
-    "last_updated": "2025-10-26T19:34:57.560324+00:00"
+    "last_updated": "2025-10-26T20:02:44.198331+00:00"
   },
   "files": [
     {
       "name": "segments.geojson",
       "present": true,
-      "modified": "2025-10-26T19:34:57.557501+00:00"
+      "modified": "2025-10-26T20:02:44.196423+00:00"
     },
     {
       "name": "segment_metrics.json",
       "present": true,
-      "modified": "2025-10-26T19:34:56.924452+00:00"
+      "modified": "2025-10-26T20:02:43.553691+00:00"
     },
     {
       "name": "flags.json",
       "present": true,
-      "modified": "2025-10-26T19:34:56.925033+00:00"
+      "modified": "2025-10-26T20:02:43.554307+00:00"
     },
     {
       "name": "meta.json",
       "present": true,
-      "modified": "2025-10-26T19:34:56.760828+00:00"
+      "modified": "2025-10-26T20:02:43.333962+00:00"
     },
     {
       "name": "flow.json",
       "present": true,
-      "modified": "2025-10-26T19:34:57.242893+00:00"
+      "modified": "2025-10-26T20:02:43.875980+00:00"
     },
     {
       "name": "schema_density.json",
       "present": true,
-      "modified": "2025-10-26T19:34:57.560280+00:00"
+      "modified": "2025-10-26T20:02:44.198294+00:00"
     }
   ],
   "hashes": {
