+++ artifacts/latest.json
@@ -1,4 +1,4 @@
 {
   "run_id": "2025-10-26",
-  "ts": "2025-10-26T19:34:00Z"
+  "ts": "2025-10-26T20:02:00Z"
 }
