+++ docs/PHASE4_RESULTS.md
@@ -1,0 +1,163 @@
+# Phase 4 Results: API Layer Refactor
+
+**Date**: 2025-10-26  
+**Issue**: #345  
+**Status**: Complete  
+**Branch**: `issue-345-phase4-api-refactor`
+
+---
+
+## Summary
+
+Successfully refactored the API layer to decouple FastAPI endpoints from core logic, introduced typed Pydantic models, and maintained 100% backward compatibility through adapter shims.
+
+---
+
+## Objectives vs. Results
+
+| Objective | Status | Details |
+|-----------|--------|---------|
+| **Decouple FastAPI endpoints from core logic** | ✅ Complete | Moved API routes to dedicated `/api/` package |
+| **Move shared business logic out of routes** | ✅ Complete | Routes now call `/core/` modules instead of inline logic |
+| **Introduce typed Pydantic models** | ✅ Complete | Created dedicated `/api/models/` package with typed schemas |
+| **Clean up late/dynamic imports** | ✅ Complete | Removed eval() calls, standardized imports |
+| **Maintain backward compatibility** | ✅ Complete | All existing imports work via adapter shims |
+
+---
+
+## New Directory Structure
+
+```
+/api/
+├── density.py          # Moved from app/density_api.py
+├── flow.py             # Moved from app/routes/api_flow.py
+├── map.py              # Moved from app/map_api.py
+├── report.py           # Moved from app/report.py
+└── models/
+    ├── density.py      # Pydantic models for density endpoints
+    ├── flow.py         # Pydantic models for flow endpoints
+    ├── map.py          # Pydantic models for map endpoints
+    └── report.py       # Pydantic models for report endpoints
+```
+
+---
+
+## Files Relocated
+
+| Original Location | New Location | Status |
+|------------------|--------------|--------|
+| `app/density_api.py` | `api/density.py` | ✅ Moved + updated imports |
+| `app/map_api.py` | `api/map.py` | ✅ Moved + updated imports |
+| `app/routes/api_flow.py` | `api/flow.py` | ✅ Moved + updated imports |
+| `app/report.py` | `api/report.py` | ✅ Moved + updated imports |
+
+---
+
+## Pydantic Models Created
+
+### Density Models (`api/models/density.py`)
+- `DensityAnalysisRequest` - Request model for density analysis
+- `DensityAnalysisResponse` - Response model for density analysis
+- `DensityReportRequest` - Request model for density report generation
+- `DensityReportResponse` - Response model for density report generation
+
+### Flow Models (`api/models/flow.py`)
+- `FlowSegmentsResponse` - Response model for flow segments endpoint
+- `FlowAnalysisRequest` - Request model for flow analysis
+- `FlowAnalysisResponse` - Response model for flow analysis
+
+### Map Models (`api/models/map.py`)
+- `MapManifestResponse` - Response model for map manifest endpoint
+- `MapBinsRequest` - Request model for map bins endpoint
+- `MapBinsResponse` - Response model for map bins endpoint
+
+### Report Models (`api/models/report.py`)
+- `ReportRequest` - Request model for report generation (with camelCase alias)
+- `ReportResponse` - Response model for report generation
+- `CombinedReportRequest` - Request model for combined report generation
+- `CombinedReportResponse` - Response model for combined report generation
+
+---
+
+## Adapter Layer Implementation
+
+All original files now contain import shims:
+
+```python
+# DEPRECATED – logic moved to api/density.py
+from api.density import *
+```
+
+This ensures:
+- ✅ **Zero breaking changes** - All existing imports continue to work
+- ✅ **Backward compatibility** - No code changes required in dependent modules
+- ✅ **Clear deprecation path** - Files marked as deprecated for future removal
+
+---
+
+## Import Updates Applied
+
+Updated imports in moved modules to use absolute paths:
+- `from .bin_analysis import` → `from app.bin_analysis import`
+- `from .constants import` → `from app.constants import`
+- `from .density import` → `from core.density.compute import`
+- `from .flow import` → `from core.flow.flow import`
+
+---
+
+## Testing Results
+
+### E2E Tests
+- ✅ **Local E2E**: All tests pass
+- ✅ **API Endpoints**: Health, density-report, temporal-flow-report all working
+- ✅ **UI Artifacts**: Generated successfully
+
+### Import Validation
+- ✅ **Main App Import**: `from app.main import app` works correctly
+- ✅ **API Endpoints**: All endpoints responding correctly
+- ✅ **Adapter Shims**: All adapter imports work correctly
+
+### Compatibility Verification
+- ✅ **Route URLs**: No changes to endpoint paths
+- ✅ **Response Formats**: All responses maintain same structure
+- ✅ **API Layer**: No changes to API-facing behavior
+
+---
+
+## Key Benefits Achieved
+
+1. **Clear Separation**: API routes separated from core business logic
+2. **Typed Interfaces**: Pydantic models provide type safety and validation
+3. **Modular Structure**: API layer organized by functionality
+4. **Maintainability**: Easier to locate and modify API-specific code
+5. **Zero Disruption**: No breaking changes to existing API consumers
+
+---
+
+## Compatibility Features
+
+### CamelCase Support
+Pydantic models include aliases for legacy camelCase inputs:
+```python
+class ReportRequest(BaseModel):
+    report_type: str = Field(..., alias="reportType")
+```
+
+### Core Module Integration
+API routes now call core modules instead of inline logic:
+```python
+from core.density.compute import analyze_density_segments
+from core.flow.flow import analyze_temporal_flow_segments
+```
+
+---
+
+## Next Steps
+
+**Phase 4 Status**: ✅ **COMPLETE**
+
+**Ready for Phase 5**: Optional cleanup phase to remove adapter shims once all dependent code is updated to use direct API imports.
+
+---
+
+**Phase 4 Resolution**: **Successfully refactored API layer with zero breaking changes and full backward compatibility maintained.**
