+++ reports/2025-10-26/bin_summary.json
@@ -1,5 +1,5 @@
 {
-  "generated_at": "2025-10-26T19:32:58.706969+00:00",
+  "generated_at": "2025-10-26T20:00:44.763556+00:00",
   "summary": {
     "total_bins": 19440,
     "flagged_bins": 1875,
